+++ src/main/java/com/codevumc/codev_backend/mapper/CoMyPageMapper.java
@@ -6,5 +6,5 @@ import org.apache.ibatis.annotations.Param;
 
 @Mapper
 public interface CoMyPageMapper {
-    boolean updateCoPortfolio(CoPortfolio coPortfolio);
+
 }
