+++ src/main/java/com/codevumc/codev_backend/mapper/CoPortfolioMapper.java
@@ -13,4 +13,5 @@ public interface CoPortfolioMapper {
     void insertCoLanguageOfPortfolio(@Param("co_portfolioId") long co_portfolioId, @Param("co_languageId") long co_languageId);
     void insertCoLinkOfPortfolio(@Param("co_portfolioId") long co_portfolioId, @Param("co_link") String co_link);
     Optional<CoPortfolio> getCoPortfolio(Map<String, Object> coPortfolioDto);
+    boolean updateCoPortfolio(CoPortfolio coPortfolio);
 }
