--- src/main/java/com/codevumc/codev_backend/service/co_mypage/CoMyPageService.java
@@ -1,9 +1,0 @@
-package com.codevumc.codev_backend.service.co_mypage;
-
-import com.codevumc.codev_backend.domain.CoPortfolio;
-import com.codevumc.codev_backend.errorhandler.CoDevResponse;
-import org.json.simple.JSONArray;
-
-public interface CoMyPageService {
-    CoDevResponse updateCoPortfolio(CoPortfolio coPortfolio, JSONArray co_languages, JSONArray co_links);
-}
