--- src/main/java/com/codevumc/codev_backend/service/co_mypage/CoMyPageServiceImpl.java
@@ -1,47 +1,0 @@
-package com.codevumc.codev_backend.service.co_mypage;
-
-import com.codevumc.codev_backend.domain.CoPortfolio;
-import com.codevumc.codev_backend.errorhandler.CoDevResponse;
-import com.codevumc.codev_backend.mapper.CoMyPageMapper;
-import com.codevumc.codev_backend.mapper.CoPortfolioMapper;
-import com.codevumc.codev_backend.service.ResponseService;
-import lombok.AllArgsConstructor;
-import org.json.simple.JSONArray;
-import org.springframework.stereotype.Service;
-
-import java.util.HashMap;
-import java.util.Map;
-import java.util.Optional;
-
-@AllArgsConstructor
-@Service
-public class CoMyPageServiceImpl extends ResponseService implements CoMyPageService{
-    private final CoMyPageMapper coMyPageMapper;
-    private final CoPortfolioMapper coPortfolioMapper;
-    @Override
-    public CoDevResponse updateCoPortfolio(CoPortfolio coPortfolio, JSONArray co_languages, JSONArray co_links) {
-        Map<String,Object> coPortfolioDto = new HashMap<>();
-        coPortfolioDto.put("co_email",coPortfolio.getCo_email());
-        coPortfolioDto.put("co_portfolioId",coPortfolio.getCo_portfolioId());
-        try{
-            Optional<CoPortfolio> coPortfolio1 = coPortfolioMapper.getCoPortfolio(coPortfolioDto);
-            if(coPortfolio1.isPresent()){
-                if(coMyPageMapper.updateCoPortfolio(coPortfolio)){
-                    for (Object co_language : co_languages) {
-                        long co_languageId = (long) co_language;
-                        coPortfolioMapper.insertCoLanguageOfPortfolio(coPortfolio.getCo_portfolioId(), co_languageId);
-                    }
-                    for (Object co_plink : co_links) {
-                        String co_link = (String) co_plink;
-                        coPortfolioMapper.insertCoLinkOfPortfolio(coPortfolio.getCo_portfolioId(), co_link);
-                    }
-                    return setResponse(200,"Complete","수정 완료되었습니다.");
-                }
-            }
-        }catch (Exception e){
-            e.printStackTrace();
-        }
-        return null;
-    }
-
-}
