+++ src/main/java/com/codevumc/codev_backend/service/co_portfolio/CoPortfolioService.java
@@ -7,4 +7,5 @@ import com.codevumc.codev_backend.errorhandler.CoDevResponse;
 public interface CoPortfolioService {
     void insertCoPortfolio(CoPortfolio coPortfolio, JSONArray co_languages, JSONArray co_links);
     CoDevResponse getCoPortfolio(long co_portfolioId, String co_email);
+    CoDevResponse updateCoPortfolio(CoPortfolio coPortfolio, JSONArray co_languages, JSONArray co_links);
 }
