--- src/main/resources/com/codevumc/codev_backend/mapper/CoMyPageMapper.xml
@@ -1,18 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" ?>
-<!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
-        "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-
-<mapper namespace="com.codevumc.codev_backend.mapper.CoMyPageMapper">
-
-    <update id="updateCoPortfolio" parameterType="CoPortfolio">
-        update CoPortfolio
-        set
-            co_title = #{co_title},
-            co_rank = #{co_rank},
-            co_headLine=#{co_headLine},
-            co_introduction =#{co_introduction}
-        where co_portfolioId = ${co_portfolioId}
-    </update>
-
-
-</mapper>
